+++ PicassoPro/Utilities/StableDiffusionAPIManager.swift
@@ -90,102 +90,3 @@ final class StableDiffusionAPIManager{
         }
     }
 }
-
-enum StableDiffusionError: Error, LocalizedError, Equatable{
-    case apiError(String)
-    case networkError(String)
-    case unknownError(String = "An  error occurred. Please retry after sometime")
-    
-    var errorDescription: String?{
-        switch self{
-        case .apiError(let message):
-            return message
-        case .networkError(let message):
-            return message
-        case .unknownError(let message):
-            return message
-        }
-    }
-}
-
-extension StableDiffusionError: RawRepresentable {
-
-    public typealias RawValue = String
-
-    public init?(rawValue: RawValue) {
-        switch rawValue {
-        case "API Error":  self = .apiError("API Error")
-        case "Network Error":  self = .networkError("Network Error")
-        case "Error":  self = .unknownError()
-        default:
-            return nil
-        }
-    }
-
-    public var rawValue: RawValue {
-        switch self {
-        case .apiError: return "API Error"
-        case .networkError: return "Network Error"
-        case .unknownError: return "Unknown Error"
-        }
-    }
-}
-
-struct ApiResponseData: Decodable{
-    var status: String
-    var generationTime: Float
-    var id: Int
-    var output: [String]
-    var meta: ApiResponseMeta
-}
-
-struct ApiResponseMeta: Decodable{
-    var H: Int
-    var W: Int
-    var enable_attention_slicing: String
-    var file_prefix: String
-    var guidance_scale: Float
-    var model: String
-    var n_samples: Int
-    var negative_prompt: String
-    var outdir: String
-    var prompt: String
-    var revision: String
-    var safetychecker: String
-    var seed: Int
-    var steps: Int
-    var vae: String
-}
-
-
-
-struct RateLimitExceededResponse: Decodable{
-    var status: String
-    var message: String
-    var tips: String
-}
-struct InvalidKeyResponse: Decodable{
-    var status: String
-    var message: String
-    var tip: String
-}
-
-struct FailedResponse: Decodable{
-    var status: String
-    var id: String
-    var message: String
-    var output: String
-}
-
-struct ValidationErrorsResponse: Decodable{
-    var status: String
-    var message: ValidationErrorsPrompt
-}
-struct ValidationErrorsPrompt: Decodable{
-    var prompt: [String]
-}
-
-struct EmptyModalIdErrorResponse: Decodable{
-    var status: String
-    var message: String
-}
